requirements.txt
absl-py==2.4.0
aioice==0.10.2
aiortc==1.14.0
altair==6.0.0
anyio==4.12.1
apsw==3.51.3.0
apswutils==0.1.2
attrs==25.4.0
av==16.1.0
base64io==1.0.3
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dnspython==2.8.0
fastcore==1.12.24
fastlite==0.2.4
flatbuffers==25.12.19
fonttools==4.62.1
gitdb==4.0.12
gitpython==3.1.46
google-crc32c==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
markupsafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
narwhals==2.18.0
numpy==2.4.3
oauthlib==3.3.1
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
pyee==13.0.1
pylibsrtp==1.0.0
pyopenssl==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-fasthtml==0.12.50
python-multipart==0.0.22
pytz==2026.1.post1
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
sounddevice==0.5.5
soupsieve==2.8.3
starlette==0.52.1
streamlit==1.55.0
streamlit-webrtc==0.64.5
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
